rename the example to hello-world
