+++ hello-world/src/main/scala/com/mesosphere/usi/helloworld/Main.scala
@@ -1,5 +1,4 @@
-package com.mesosphere.usi.example
-
+package com.mesosphere.usi.helloworld
 
 import com.mesosphere.usi.core.Scheduler
 import com.mesosphere.usi.interface.PodSpec
